refactor: merge read_sql_utils into sql_utils
